Fixed App test classpath after rpsl4j dependency added and added new process timeout code.
The AppTest class previously depended on the ripe-rpsl code being in the same repository as itself.
This requirement was removed by passing the classpath of the testing JVM to the spawned one.
The process timeout code was also rewritten such that early terminating processes are detected before the max timeout