Add BGPAutNum and BGPRoute, deprecated ODLConfigEmitter
Due to expected issues in the implementation of routing in the previous
ODLConfigEmitter classes, they have been deprecated and work on a new
implementation. Currently BGPAutNum builds tables of routes to export to
specific peers in remote AS's. BGPRoute is responsible for parsing the
routes, but doesn't presently support filter expressions (AND/OR/NOT).