Fixed AttributeLexerWrapper skippig opening tokens
The class was originally written to skip text captured in states until it
encountered the first keyword. This was so the attribtue key (ie import:)
could be skipped. This meant that attributes, such as ifaddr, which open
with a token group (ifaddr: 1.2.3.4) would not capture this group. This
has been fixed and tests added.