Implemented BGPPeer and BGPInetRtr (speaker)
BGP speaker and peer support have been added in the forms of BGPInetRtr
and BGPPeer respectively. InetRtrs, when constructed, instantiate the
peers declared in their rpsl-object. These peers in turn look up their
AS in the AutNum of the InetRtr and build their route sets. Method to
lookup these AS's was added to BGPAutNum.