Added initial support for AS routes and route-sets
BGPRpslDocument now allows for route objects with an AS origin to be
queried. BGPRoute#resolveRoutes will check, if provided an RpslDocument,
if a route string is infact an AS, and will add routes originating from
the AS if so.

Scaffold code for building and resolving route/as-set objects has been
added to BGPRpslDocument and BGPRpslSet.