Split BGPRpslRoute out of BGPRoute
fields that only made sence for actual rpsl-route objects were being
added to BGPRoute (which is used as route prefix + nexthop). The class
was specialised for the case where a route is built from an RpslObject
and the copy constructor was replaced with clone to allow generic
copying of all route objects.